plcc

+++ CompressionForce.Domain/PLC/ServoSetDto.cs
@@ -5,15 +5,15 @@ using System.Text;
 using System.Threading.Tasks;
 
 
-    namespace CompressionForce.Domain.PLC
-    {
+namespace CompressionForce.Domain.PLC
+{
     public class ServoSetDto
     {
         public string ServoCode { get; set; }
-        public double SetValue { get; set; }     
-        public double SetPosition { get; set; }
-        public double SetSpeed { get; set; }
-        public double JogSpeed { get; set; }
+        public int SetValue { get; set; }
+        public int SetPosition { get; set; }
+        public int SetSpeed { get; set; }
+        public int JogSpeed { get; set; }
     }
 
 }
